clean up code
- remove unused methods and attributes

- broke up classes into multiple files to improve readability
