+++ src/main/java/Dude/DialogBox.java
@@ -39,19 +39,6 @@ public class DialogBox extends HBox {
         displayPicture.setImage(img);
     }
 
-    private Label text;
-
-    public DialogBox(Label l, ImageView iv) {
-        text = l;
-        displayPicture = iv;
-
-        text.setWrapText(true);
-        displayPicture.setFitWidth(100.0);
-        displayPicture.setFitHeight(100.0);
-
-        this.setAlignment(Pos.TOP_RIGHT);
-        this.getChildren().addAll(text, displayPicture);
-    }
 
     /**
      * Flips the dialog box such that the ImageView is on the left and text on the right.
@@ -73,13 +60,4 @@ public class DialogBox extends HBox {
         return db;
     }
 
-    public static DialogBox getUserDialog(Label l, ImageView iv) {
-        return new DialogBox(l, iv);
-    }
-
-    public static DialogBox getDudeDialog(Label l, ImageView iv) {
-        var db = new DialogBox(l, iv);
-        db.flip();
-        return db;
-    }
 }
